feat(user management): user already exists, password hash, cancel buttons

+++ public/index.php
@@ -36,21 +36,27 @@ if (isset($_POST) and !empty($_POST)) {
             $adminCommentsController->addComment($_POST);
             break;
         case "add_user":
-            if (
-                isset($_FILES['avatar'])
-                and $_FILES['avatar']['error'] == 0
-                and ($_FILES['avatar']['size'] <= 1000000)
-            ) {
-                $infoFileUploaded = pathinfo($_FILES['avatar']['name']);
-                $extensionFileUploaded = $infoFileUploaded['extension'];
-                $authorizedExtensions = array('jpg', 'jpeg', 'gif', 'png');
-                if (in_array($extensionFileUploaded, $authorizedExtensions)) {
-                    move_uploaded_file($_FILES['avatar']['tmp_name'], './img/' . basename($_FILES['avatar']['name']));
-                }
-                $_POST["avatar"] = $_FILES['avatar']['name'];
-            }
             $userController = new \App\controller\UserController();
-            $userController->addUser($_POST);
+            $email = $_POST['email'];
+            if (!$userController->userExist($email)) {
+                if (
+                    isset($_FILES['avatar'])
+                    and $_FILES['avatar']['error'] == 0
+                    and ($_FILES['avatar']['size'] <= 1000000)
+                ) {
+                    $infoFileUploaded = pathinfo($_FILES['avatar']['name']);
+                    $extensionFileUploaded = $infoFileUploaded['extension'];
+                    $authorizedExtensions = array('jpg', 'jpeg', 'gif', 'png');
+                    if (in_array($extensionFileUploaded, $authorizedExtensions)) {
+                        move_uploaded_file($_FILES['avatar']['tmp_name'], './img/' . basename($_FILES['avatar']['name']));
+                    }
+                    $_POST["avatar"] = $_FILES['avatar']['name'];
+                }
+                $userController->addUser($_POST);
+            } else {
+                $connectionController = new \App\controller\ConnectionController();
+                $connectionController->index(true);
+            }
             break;
         case "contact_form":
             $HomeController = new \App\controller\HomeController();
@@ -110,7 +116,7 @@ $method = isset($_GET["action"]) ? $_GET["action"] : 'index';
 if (isset($_GET['page']) && !empty($_GET['page'])) {
     $currentPage = (int)strip_tags($_GET['page']);
 } elseif ($p == 'post' or $p == 'adminPosts' or $p == 'adminComments'
-    or $p =='connection' or $p == 'adminModifyPost' && $method == 'index') {
+    or $p == 'adminModifyPost' && $method == 'index') {
     $currentPage = 1;
 }
